+++ packages/blackbox_core/test/pipeline_test.dart
@@ -56,7 +56,7 @@ void main() {
       final pipeline = PipelineBuilder()
           .add(a)
           .add(b)
-          .addWithDependencies(
+          .addWith(
             sum,
             dependencies: (d) => (
               a: d.require<int>(a),