+++ packages/blackbox_flutter/lib/blackbox_flutter.dart
@@ -5,6 +5,7 @@
 library blackbox_flutter;
 
 import 'package:blackbox/blackbox.dart';
+import 'package:meta/meta.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/widgets.dart';
 
@@ -13,3 +14,4 @@ export 'src/persistent_shared_prefs_store.dart';
 part 'src/_tracker.dart';
 part 'src/box_provider.dart';
 part 'src/box_observer.dart';
+part 'src/observable_flow_box.dart';